Sea: jiggle modules for Zebra
